+++ zanshincli/src/dummy_aws_credentials
