+++ zanshincli/src/version.py
